added Talking With AI feature and improvements fix conflicting contents
