+++ .gitignore
@@ -7,5 +7,6 @@ manifest.ini
 *.pot
 *.py[co]
 *.nvda-addon
+*.back
 .sconsign.dblite
 /[0-9]*.[0-9]*.[0-9]*.json
